fix: 연구실 fetch 수정
