+++ app/[locale]/people/faculty/[id]/edit/page.tsx
@@ -1,4 +1,5 @@
 import { getFaculty } from '@/apis/people';
+import { getResearchLabs } from '@/apis/research';
 
 import { Language } from '@/types/language';
 
@@ -14,8 +15,20 @@ export default async function FacultyEditPage({
   const id = +rawId;
   if (Number.isNaN(id)) throw new Error('유효한 id가 아닙니다: ' + rawId);
 
-  const data = await getFaculty(id);
+  // const data = await getFaculty(id);
+  const [faculty, koLabs, enLabs] = await Promise.all([
+    getFaculty(id),
+    getResearchLabs('ko'),
+    getResearchLabs('en'),
+  ]);
 
   // TODO: 영어 데이터 같이 오는 형식 확정되면 해당 사항 반영해서 data 타입 수정
-  return <FacultyEditPageContent data={data} language={locale} id={{ ko: id, en: id }} />;
+  return (
+    <FacultyEditPageContent
+      data={faculty}
+      labs={{ ko: koLabs, en: enLabs }}
+      language={locale}
+      id={{ ko: id, en: id }}
+    />
+  );
 }
